+++ src/proto/mod.rs
@@ -1,9 +1,12 @@
 use std::io::Write;
 use std::net::TcpStream;
 
-use crate::core::proto::{Message, Type};
 use crate::error::{Error, ErrCode, convert_err};
 
+pub mod message;
+use message::{Message, Type};
+
+/// Write specified message into the stream
 pub fn send(stream: &mut TcpStream, message: Message) -> Result<(), Error> {
     stream.write_all(
         message
@@ -18,4 +21,16 @@ pub fn handshake_init(stream: &mut TcpStream) -> Result<bool, Error> {
     send(stream, Message::new_request())?;
     let reply = Message::deserialize(stream)?;
     Ok(if reply.t == Type::Accept {true} else {false})
+}
+
+
+/// Try recieveng message; if it's a valid request,
+/// a decline message is sent back.
+pub fn decline(mut stream: TcpStream) {
+    // TODO error handling
+    if let Ok(msg) = Message::deserialize(&mut stream) {
+        if msg.t == Type::Request {
+            send(&mut stream, Message::new_deny()).unwrap();
+        }
+    }
 }
